+++ Changelog.md
@@ -5,6 +5,9 @@ v1.0.4 Unreleased
 * Propagate uncertainties from radiometry to OC products
 * Revisit question of overly aggressive deglitching
 
+2020-11-27:
+* Add Wei et al. 2016 spectral QA score to L2 products. Include on Rrs plots.
+
 2020-11-20:
 * Fix NIR dataset for when no NIR residual correction is performed.